+++ packages/core/src/modules/wms/data/validators.ts
@@ -178,8 +178,8 @@ export const inventoryLotUpdateSchema = z
 const inventoryLowStockFilterSchema = z.enum(['belowReorder', 'belowSafety'])
 
 export const inventoryBalanceListQuerySchema = z.object({
-  page: z.coerce.number().min(1).default(1),
-  pageSize: z.coerce.number().min(1).max(100).default(25),
+  page: z.coerce.number().int().min(1).default(1),
+  pageSize: z.coerce.number().int().min(1).max(100).default(25),
   warehouseId: uuid().optional(),
   locationId: uuid().optional(),
   catalogVariantId: uuid().optional(),
@@ -190,8 +190,8 @@ export const inventoryBalanceListQuerySchema = z.object({
 }).passthrough()
 
 export const inventoryMovementListQuerySchema = z.object({
-  page: z.coerce.number().min(1).default(1),
-  pageSize: z.coerce.number().min(1).max(100).default(25),
+  page: z.coerce.number().int().min(1).default(1),
+  pageSize: z.coerce.number().int().min(1).max(100).default(25),
   warehouseId: uuid().optional(),
   locationId: uuid().optional(),
   catalogVariantId: uuid().optional(),
@@ -203,8 +203,8 @@ export const inventoryMovementListQuerySchema = z.object({
 }).passthrough()
 
 export const inventoryReservationListQuerySchema = z.object({
-  page: z.coerce.number().min(1).default(1),
-  pageSize: z.coerce.number().min(1).max(100).default(25),
+  page: z.coerce.number().int().min(1).default(1),
+  pageSize: z.coerce.number().int().min(1).max(100).default(25),
   warehouseId: uuid().optional(),
   catalogVariantId: uuid().optional(),
   lotId: uuid().optional(),
